Finish Information Section Styling
Information section styling complete, needed to use two different
flex directions when organizing the subsections and the actual
"images". Also added some of the formatting of the colors, and
sizes of the text on the web page. Figured out how to import
different font from google fonts and using them in my stylesheet.