Project Finish and Finishing Touches
Finished footer with and properly corresponding links.
Also added some touches to make the website look nicer in
general such as rounding some corners and adding a funny
image to go with the quote. Links were all wrong and had
to be redone with new href links. Also gained a better
understanding of aligning/centering items. Also added the
image to the media directory.